requirements.txt
redis>=5.0.0
PyYAML>=6.0
pydantic>=2.7
